making some changes

+++ src/engine/gridobject/Door.java
@@ -19,7 +19,7 @@ public class Door extends GridObject{
 	}
 	
 	public Door(List<Object> list){
-		super((String) list.get(Constants.IMAGE_CONST), (int) list.get(Constants.WIDTH_CONST), (int) list.get(Constants.HEIGHT_CONST));
+		super((String) list.get(Constants.IMAGE_CONST), (int) ((Double) list.get(Constants.WIDTH_CONST)).intValue(), (int) ((Double) list.get(Constants.HEIGHT_CONST)).intValue());
 	}
 	
 	/**
